% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcauchyd-package.R
\docType{package}
\name{mcauchyd-package}
\alias{mcauchyd-package}
\alias{mcauchyd}
\title{Tools for Multivariate Cauchy Distributions}
\description{
This package provides tools for multivariate Cauchy distributions (MCD):
\itemize{
\item Calculation of distances/divergences between MCD:
\itemize{
\item Kullback-Leibler divergence: \code{\link{kldcauchy}}
}
\item Tools for MCD:
\itemize{
\item Probability density: \code{\link{dmcd}}
\item Simulation from a MCD: \code{\link{rmcd}}
\item Plot of the density of a MCD with 2 variables: \code{\link{plotmcd}}, \code{\link{contourmcd}}
}
}
}
\references{
N. Bouhlel, D. Rousseau, A Generic Formula and Some Special Cases for the Kullback–Leibler Divergence between Central Multivariate Cauchy Distributions.
Entropy, 24, 838, July 2022.
\doi{10.3390/e24060838}
#' @keywords internal
}
\seealso{
Useful links:
\itemize{
  \item \url{https://forge.inrae.fr/imhorphen/mcauchyd}
  \item Report bugs at \url{https://forge.inrae.fr/imhorphen/mcauchyd/-/boards}
}

}
\author{
Pierre Santagostini \href{mailto:pierre.santagostini@agrocampus-ouest.fr}{pierre.santagostini@agrocampus-ouest.fr},
Nizar Bouhlel \href{mailto:nizar.bouhlel@agrocampus-ouest.fr}{nizar.bouhlel@agrocampus-ouest.fr}
}
